requirements.txt
streamlit==1.38.0
groq>=0.11.0
python-dotenv>=1.0.1
